requirements.txt
discord.py==1.0.1
Django==2.0
django-allauth==0.38.0
dj-database-url==0.5.0
gunicorn==19.8.1
requests==2.18.4
requests-oauthlib==0.8.0
psycopg2==2.7.7
python-dateutil==2.7.2
whitenoise==3.3.1
